create @app/shared.
- seems like working now
