+++ packages/data/package.json
@@ -9,7 +9,7 @@
     "update-data:skip-commit": "npm run update-data -- skip-commit"
   },
   "devDependencies": {
-    "@app/type": "workspace:*",
+    "@app/shared": "workspace:*",
     "@supercharge/promise-pool": "^2.3.2",
     "@types/node": "^18.6.4",
     "ava": "^4.3.1",
