+++ packages/extension/package.json
@@ -34,7 +34,7 @@
     "safe-compare": "1.1.4"
   },
   "devDependencies": {
-    "@app/type": "workspace:*",
+    "@app/shared": "workspace:*",
     "@babel/core": "^7.18.10",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-proposal-object-rest-spread": "^7.18.9",
