--- packages/type/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "@app/type",
-  "version": "1.0.0",
-  "private": true,
-  "description": "This is to host Tradingview interfaces & types",
-  "license": "MIT"
-}
