+++ packages/web/.env.example
@@ -1,2 +1,2 @@
 VITE_GA=GA
-VITE_FETCH_URL=https://raw.githubusercontent.com/AzrizHaziq/tradingview-syariah-indicator/master/packages/data/stock-list-human.json
+VITE_FETCH_URL=https://raw.githubusercontent.com/AzrizHaziq/tradingview-syariah-indicator/master/packages/data
