+++ packages/web/src/routes/list/index.tsx
@@ -1,9 +1,10 @@
 import './list.scss'
+import { debounce } from '@app/shared'
 import { MetaSeo } from '~/components'
 import { Head, Title } from 'solid-start'
-import { ShariahFetcher, PageProps } from './fetcher'
 import { createStore } from 'solid-js/store'
-import { debounce, IMGS, pipe, TArrayConcat, TFilter, useTrackOnLoad } from '~/util'
+import { ShariahFetcher, PageProps } from './fetcher'
+import { IMGS, pipe, TArrayConcat, TFilter, useTrackOnLoad } from '~/util'
 import { createEffect, createMemo, createResource, For, JSX, Show } from 'solid-js'
 
 const staticExchangeColors = [
