+++ packages/web/src/util/functions.ts
@@ -1,13 +1,3 @@
-export function debounce(func, wait = 300) {
-  let timeout
-  return function (...args) {
-    clearTimeout(timeout)
-    timeout = setTimeout(() => {
-      func.apply(this, args)
-    }, wait)
-  }
-}
-
 export const pipe =
   (...fns) =>
   (x) =>
